almost compleompleted buffeditor, now need the consumer to add the list to Collider

+++ src/engine/gameobject/units/Buff.java
@@ -65,7 +65,7 @@ public abstract class Buff implements Upgrade {
      *
      * @return
      */
-    public int timeLeft () {
+    public double timeLeft () {
         return getDuration() - timeSinceStart;
     }
